✨ 디렉토리위치수정(history>history/tab) &현재 클릭한 탭 인식 기능 추가
